requirements.txt
Django==4.2
djangorestframework==3.14.0
xlwt==1.3.0
redis==5.0.7
celery==5.4.0
requests==2.31.0
